Consolidate question-timeline UI into query_builder (DEV-2340)
Move query_builder-only code to its consumer:

- timelines/questions/ (components + containers: EventCard, TimelineCard,
  TimelineEmptyState, TimelineList, TimelinePanel, Edit/Move/NewEventModal)
  -> query_builder/components/timelines/. This subtree is the question-view
  timeline sidebar; its only consumers are TimelineSidebar and QueryModals,
  and it has zero imports from the rest of shared/timelines.
- common/components/ViewFooterButton -> query_builder/components/view/.
  It hardcodes the qb-specific viewFooter ActionIcon variant and is used
  only by QuestionTimelineWidget.

Verified via dependency-graph analysis + repo-wide greps (incl. e2e).
Pure git mv + import updates; no behavior change.
