+++ frontend/src/metabase/query_builder/components/timelines/components/TimelineEmptyState/index.ts
