+++ frontend/src/metabase/query_builder/components/timelines/containers/NewEventModal/NewEventModal.tsx
