+++ frontend/src/metabase/query_builder/components/view/ViewFooterButton/index.ts
